requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.5.0
